cli: decode every PVE scalar through the tolerant SDK types
PVE renders a Perl scalar as a JSON string or a JSON number depending on
how that SV was last used, so a field documented as an integer arrives as
5 on one node and "5" on the next. A rigid Go int or bool fails the whole
decode the first time the other form shows up, which is how `lab ceph
osd` died on `cannot unmarshal string into ... osdid of type int`.

Convert the remaining 81 rigid scalars in PVE-facing decoders to
client.PVEInt and client.PVEFloat, and add an AST test that walks every
struct a json.Unmarshal decodes into and fails on a rigid scalar. PBS and
PDM are exempt: they are Rust behind serde, so a documented integer is
always a number. Deliberate exceptions go in decodeAuditAllowed with a
reason.

PVEInt and PVEFloat marshal back as JSON numbers, so -o json output is
unchanged.

+++ internal/cli/lab/scale.go
@@ -427,7 +427,7 @@ func scaleRenameLegacyNodeZero(ctx context.Context, deps *cli.Deps, lab *config.
 		return "", err
 	}
 
-	vmidStr := strconv.FormatInt(vm.VMID, 10)
+	vmidStr := strconv.FormatInt(int64(vm.VMID), 10)
 	if err := deps.API.Nodes.UpdateQemuConfig(ctx, vm.Node, vmidStr, &nodes.UpdateQemuConfigParams{Name: new(newName)}); err != nil {
 		return "", fmt.Errorf("rename legacy VM %d (%q) to %q on node %q: %w", vm.VMID, legacyName, newName, vm.Node, err)
 	}
